+++ testing/sync-test/src/logins.rs
@@ -10,7 +10,6 @@ use logins::{
 };
 use std::sync::Arc;
 use std::{collections::hash_map::RandomState, collections::HashMap};
-use sync15::ServerTimestamp;
 
 // helpers...
 // Doesn't check metadata fields
@@ -339,94 +338,10 @@ fn test_login_deletes(c0: &mut TestClient, c1: &mut TestClient) {
     verify_missing_login(&c0.logins_store, &l2id);
 }
 
-// This test verifies that the delete_local function deletes a login without propogating the
-// deletion to the server
-fn test_login_local_delete(c0: &mut TestClient, c1: &mut TestClient) {
-    log::info!("Add a login to client0");
-
-    // Add a login
-    let login0 = add_login(
-        &c0.logins_store,
-        LoginEntry {
-            origin: "http://www.example2.com".into(),
-            form_action_origin: Some("http://login.example2.com".into()),
-            username_field: "uname".into(),
-            password_field: "pword".into(),
-            username: "cool_username".into(),
-            password: "hunter2".into(),
-            ..Default::default()
-        },
-    )
-    .expect("add l0");
-
-    let l0id = login0.guid();
-
-    // Sync the first device where the login was added
-    log::info!("Syncing client0 -- inital sync");
-    sync_logins(c0).expect("c0 sync to work");
-
-    // Sync the second device
-    log::info!("Syncing client1 -- inital sync");
-    sync_logins(c1).expect("c0 sync to work");
-
-    // Verify that the login exists on both devices
-    verify_login(&c0.logins_store, &login0);
-    verify_login(&c1.logins_store, &login0);
-
-    // Delete the login on the first device
-    log::info!("Performing local delete {} from c0", l0id);
-    c0.logins_store
-        .db
-        .lock()
-        .delete_local_records_for_remote_replacement(vec![&l0id])
-        .expect("Local delete should work");
-
-    // Verify that the login no longer exists on the first device
-    verify_missing_login(&c0.logins_store, &l0id);
-
-    // Sync the first device
-    log::info!("Syncing client0 -- post local deletion");
-    sync_logins(c0).expect("c0 sync to work");
-
-    // Sync the second device
-    log::info!("Syncing client1 -- post local deletion");
-    sync_logins(c1).expect("c1 sync to work");
-
-    // The following verification checks ensure that the delete_local_for_remote_replacement
-    // call removed the login from the first device without propogating that change to the
-    // sync server.
-
-    // Verify that the login is missing from the first device.
-    verify_missing_login(&c0.logins_store, &l0id);
-
-    // Verify that the login still exists on the second device
-    verify_login(&c1.logins_store, &login0);
-
-    // Reset the last sync date for the logins sync engine. This will allow the previously synced
-    // and then locally deleted record to be downloaded to the first client.
-    c0.logins_store
-        .clone()
-        .set_last_sync(ServerTimestamp(0))
-        .expect("last sync to be set");
-
-    // Sync the first device
-    log::info!("Syncing client0 -- post c0 reset");
-    sync_logins(c0).expect("c0 sync to work");
-
-    // Sync the second device
-    log::info!("Syncing client1 -- post c0 reset");
-    sync_logins(c1).expect("c1 sync to work");
-
-    // Verify that the login exists on both devices
-    verify_login(&c0.logins_store, &login0);
-    verify_login(&c1.logins_store, &login0);
-
-    // Clear the stores
-    _ = c0.logins_store.wipe_local();
-    _ = c1.logins_store.wipe_local();
-}
-
-fn test_verify_logins_unblocks_syncing(c0: &mut TestClient, c1: &mut TestClient) {
+fn test_delete_undecryptable_records_for_remote_replacement(
+    c0: &mut TestClient,
+    c1: &mut TestClient,
+) {
     log::info!("Add a login to client0");
 
     // Add a login
@@ -506,11 +421,10 @@ fn test_verify_logins_unblocks_syncing(c0: &mut TestClient, c1: &mut TestClient)
 
     // Execute the verification logic to remove the corrupted login
     log::info!("Verify logins");
-    assert!(c0
-        .logins_store
+    c0.logins_store
         .clone()
-        .verify_logins()
-        .expect("stored logins to be verified"));
+        .delete_undecryptable_records_for_remote_replacement()
+        .expect("stored logins to be verified");
 
     // Verify that the corrupted login has been removed
     verify_missing_login(&c0.logins_store, &l1id);
@@ -538,10 +452,9 @@ pub fn get_test_group() -> TestGroup {
         vec![
             ("test_login_general", test_login_general),
             ("test_login_deletes", test_login_deletes),
-            ("test_login_local_delete", test_login_local_delete),
             (
-                "test_verify_logins_unblocks_syncing",
-                test_verify_logins_unblocks_syncing,
+                "test_delete_undecryptable_records_for_remote_replacement",
+                test_delete_undecryptable_records_for_remote_replacement,
             ),
         ],
     )